+++ src/components/Alerts/AlertSystem.js
@@ -5,7 +5,7 @@ import Alert from './Alert'
 
 class AlertSystem extends React.Component {
   constructor(props) {
-    super(props);
+    super(props)
   }
 
   state = {
@@ -21,40 +21,46 @@ class AlertSystem extends React.Component {
      * Optional styles to add to the alert system component.
      */
     optClass: React.PropTypes.string
-  };
+  }
 
   getAlerts = () => {
     return this.state.alerts.map((alert, index) =>
-      !alert.hidden ? <Alert key={alert.key} type={alert.type || 'success'} optClass={alert.class || ''} closable={typeof alert.closable !== 'undefined' ? alert.closable : true} timeout={alert.timeout} onClose={this.removeAlert.bind(this, alert.key)}>{alert.content}</Alert> : null
-    );
+      !alert.hidden ? <Alert key={alert.key} type={alert.type || 'success'} optClass={alert.class || ''} closable={typeof alert.closable !== 'undefined' ? alert.closable : true} timeout={alert.timeout} onClose={this.removeAlert.bind(this, alert)}>{alert.content}</Alert> : null
+    )
   }
 
-  removeAlert = (key) => {
-    let alerts = this.state.alerts;
+  removeAlert = (alert) => {
+    let alerts = this.state.alerts
 
-    alerts.map((alert, index) => {
-      if (alert.key === key) {
-        alert.hidden = true;
+    alerts.map((a, index) => {
+      if (alert.key === a.key) {
+        a.hidden = true
+
+        if (typeof a.onClose === 'function') {
+          a.onClose(alert)
+        }
       }
-    });
+    })
 
-    this.setState({ alerts: alerts });
+
+
+    this.setState({ alerts: alerts })
   }
 
   componentWillReceiveProps = (nextProps) => {
-    let alerts = nextProps.alerts;
+    let alerts = nextProps.alerts
     alerts.map((alert, index) => {
       if (!alert.key) {
-        alert.key = (alert.type || 'success') + '-' + new Date().getTime();
+        alert.key = (alert.type || 'success') + '-' + new Date().getTime()
       }
-    });
+    })
 
-    this.setState({ alerts: alerts });
+    this.setState({ alerts: alerts })
   }
 
   render() {
-    const cx = classNames.bind(style);
-    const alertSystemClasses = cx(style['alert-system'], this.props.optClass);
+    const cx = classNames.bind(style)
+    const alertSystemClasses = cx(style['alert-system'], this.props.optClass)
 
     return (
       <div className={alertSystemClasses}>